Create the getComputerChoice function, getPlayerChoice function, playRound function
getComputerChoice function generate randomly 1, 2 or 3 using Math.floor
and Math.random methods, then return Rock if 1 was generated, Paper if 2
was generated, and Scissors if 3 was generated.

getPlayerChoice function take a string as an argument, turn this
string to lower case using toLowerCse function, and then capitalize the
first letter of the string using charAt and toUppercase functions.

playRound function take the returned value of the two previous function
as arguments, and  declare the winner of the round according to the game
rules (rock, papper, scissors).